fix: refactor, check step util move to common
